// types.ts
export interface Activity {
    id: string;
    title: string;
    location: string;
    price: number;
    rating: number;
    reviews: string;
    image: string;
    category: string;
    duration: string;
    tags: string[];
}